ProjectExplorer: Add setting to not auto-remove run configurations
This change introduces a new boolean config in the ProjectExplorer
plugin. Next to the existing "Create suitable run configurations
automatically" we add "Remove obsolete run configurations
automatically", which is enabled by default.

Disabling the new setting prevents run configurations from being
removed automatically, even if the corresponding cmake target is not
present any more. This can, for example, be relevant in projects with
cmake arguments that control the targets created in the current
build. Or when switching between branches with different sets of
targets.

Without this, run configurations and their
customizations (e.g. command line arguments) are lost easily without
warning and way to prevent it.

Fixes: QTCREATORBUG-28273
Change-Id: I52c3b54be4cf047b6467d5f8e429d66d5b474cef