ProjectExplorer: Do not auto-remove run configurations
With this change, if "Create suitable run configurations
automatically" is enabled, we don't remove run configurations, even if
the corresponding target is (currently) missing.

This prevents run configurations that were likely added manually from
being removed automatically. This can be relevant in projects with
cmake arguments that control the targets created in the current
build. Or when switching between branches with different sets of
targets.

Fixes: QTCREATORBUG-?????
Change-Id: I52c3b54be4cf047b6467d5f8e429d66d5b474cef